Move bpi subject role to separate file
