+++ examples/bri-3/src/bri/identity/bpiSubjects/agents/bpiSubjects.agent.ts
@@ -4,13 +4,14 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { v4 } from 'uuid';
-import { BpiSubject, BpiSubjectRoleName } from '../models/bpiSubject';
+import { BpiSubject } from '../models/bpiSubject';
 
 import {
   NAME_EMPTY_ERR_MESSAGE,
   NOT_FOUND_ERR_MESSAGE,
 } from '../api/err.messages';
 import { BpiSubjectStorageAgent } from './bpiSubjectsStorage.agent';
+import { BpiSubjectRoleName } from '../models/bpiSubjectRole';
 
 // Agent methods have extremely declarative names and perform a single task
 @Injectable()
