+++ examples/bri-3/src/bri/identity/bpiSubjects/models/bpiSubject.ts
@@ -1,27 +1,6 @@
 import { AutoMap } from '@automapper/classes';
 import { v4 } from 'uuid';
-
-export enum BpiSubjectRoleName {
-  INTERNAL_BPI_SUBJECT = 'INTERNAL_BPI_SUBJECT',
-  EXTERNAL_BPI_SUBJECT = 'EXTERNAL_BPI_SUBJECT',
-}
-
-export class BpiSubjectRole {
-  @AutoMap()
-  id: string; // TODO: Add uuid after #491
-
-  @AutoMap()
-  name: BpiSubjectRoleName;
-
-  @AutoMap()
-  description: string;
-
-  constructor(id: string, name: BpiSubjectRoleName, description: string) {
-    this.id = id;
-    this.name = name;
-    this.description = description;
-  }
-}
+import { BpiSubjectRole } from './bpiSubjectRole';
 
 export class BpiSubject {
   @AutoMap()
